Split off githubdiff
